Inspector: re-implement placing vertical rules in probe group data grids when the main frame changes.
Separators are now implemented as a flag on the preceding data frame, which
causes a vertical rule to be rendered by the DataGrid subclass.

This is possible due to the simplifying assumption that vertical rules can only
appear after a frame (not before the first frame), and that multiple adjacent
reloads are represented by a single vertical rule.