Web Replay: ensure that DOM events are prefixed by an event loop input.
Closes #385.

A DOM event should always be causally related to an event loop input that was
captured or is being replayed. If it is not, then the DOM event could have been
dispatched nondeterministically, as we did not account for it's causal source.

In practice, this invariant boils down to whether a DOM event is dispatched while
an input is being captured or replayed earlier in the call stack.

This patch adds a simple assertion for this invariant during replay. To assert
the invariant during capturing, we model the effective 'extent' of a captured
event loop input by using the EventLoopInputExtent RAII helper class.