Web Replay: reduce edit overhead of adding new inputs.
We now use macrology to generate most of the encode method
disambiguation cases in JSONEncoderContext. This piggybacks on
the macro used by ReplayInputTypes, but it had to be modified
somewhat to only enumerate non-template classes.

To make disambiguation work with AutoMemoized specializations, we
create the AutoMemoizedBase class. It has a virtual encode()
method that calls the right specialized InputCoder<T>::encode method.

Lastly, header includes for all input types was moved to
AllInputTypes.h since header includes can't be generated by the
preprocessor easily.