Web Replay: improve capture and replay of page closing.  Closes #389.
WebPage now goes through NavigationProxy when trying to close an
entire page. This allows distinguishing between deterministic and
nondeterministic page closes, and tracking of any subsequent
beforeunload DOM events.

If the embedder wants to close the page when we are capturing,
then we end capturing unconditionally if shouldClose() returns
true. If it returns false, then we capture this unsuccessful
close as an event loop input to be simulated later.

If the embedder wants to close the page when we are replaying,
then we cancel playback and close the page if the page close
request is not being simulated (i.e., the live user requested it
during the replay, not the simulated user during the
capture). Otherwise, we continue replay and follow the normal
steps to determine whether to close the page or not.