Probes: fix breakpoint mode inconsistency when creating from popup menu.
Fixes #419.

This fixes two things:

The Breakpoint.mode setter didn't check the validity of the supplied mode.
Now, if you try to set auto-continue when no actions are attached, it will
instead set the mode to enabled.

There was a race between setting the probe breakpoint action and setting
the breakpoint's mode to auto-continue. A short timer was added as a
temporary workaround until Breakpoints have an API for batching mutations.