Probes: move the new probe expression entry popup to gutter context menu.
Fixes #309.

We used to want the quick-add for probe expressions to be in the gutter,
but CodeMirror bugs prevented that from working. So instead it was added
as a shift-click gutter action. Now that the gutter has a context menu,
the quick-add dialog is triggered from the "Add Probe Expression" context
menu item.